scriptcord command module base class
